Move project related classes to same package
