+++ common/app/com/commercetools/sunrise/ctp/project/LocalizationModule.java
@@ -1,0 +1,24 @@
+package com.commercetools.sunrise.ctp.project;
+
+import com.commercetools.sunrise.framework.injection.RequestScoped;
+import com.google.inject.AbstractModule;
+import com.neovisionaries.i18n.CountryCode;
+
+import javax.money.CurrencyUnit;
+import java.util.Locale;
+
+public final class LocalizationModule extends AbstractModule {
+
+    @Override
+    protected void configure() {
+        bind(Locale.class)
+                .toProvider(LocaleFromUrlProvider.class)
+                .in(RequestScoped.class);
+        bind(CountryCode.class)
+                .toProvider(CountryFromSessionProvider.class)
+                .in(RequestScoped.class);
+        bind(CurrencyUnit.class)
+                .toProvider(CurrencyFromCountryProvider.class)
+                .in(RequestScoped.class);
+    }
+}
