+++ common/app/com/commercetools/sunrise/ctp/project/NoCountryFoundException.java
@@ -1,4 +1,4 @@
-package com.commercetools.sunrise.framework.localization;
+package com.commercetools.sunrise.ctp.project;
 
 public class NoCountryFoundException extends RuntimeException {
