+++ common/app/com/commercetools/sunrise/ctp/project/NoCurrencyFoundException.java
@@ -1,4 +1,4 @@
-package com.commercetools.sunrise.framework.localization;
+package com.commercetools.sunrise.ctp.project;
 
 public class NoCurrencyFoundException extends RuntimeException {
