+++ common/app/com/commercetools/sunrise/ctp/project/NoLocaleFoundException.java
@@ -1,4 +1,4 @@
-package com.commercetools.sunrise.framework.localization;
+package com.commercetools.sunrise.ctp.project;
 
 public class NoLocaleFoundException extends RuntimeException {
