+++ common/app/com/commercetools/sunrise/ctp/project/ProjectContext.java
@@ -1,4 +1,4 @@
-package com.commercetools.sunrise.framework.localization;
+package com.commercetools.sunrise.ctp.project;
 
 import com.google.inject.ImplementedBy;
 import com.neovisionaries.i18n.CountryCode;
