+++ common/conf/reference.conf
@@ -9,7 +9,7 @@ sunrise {
     # CLIENT
     # ~~~~~
 
-    client = {
+    client {
 
       # Project key of your commercetools project.
       # projectKey = "your-project-key"
@@ -35,7 +35,7 @@ sunrise {
     # CATEGORIES
     # ~~~~~
 
-    categories = {
+    categories {
 
       # Key from the cache where to store the category tree.
       cacheKey = "category-tree"
@@ -69,7 +69,7 @@ sunrise {
     # PRODUCTS
     # ~~~~~
 
-    products = {
+    products {
 
       attributes {
 
@@ -87,25 +87,24 @@ sunrise {
         secondarySelectable = []
       }
     }
-  }
 
+    # PROJECT
+    # ~~~~~
 
-  # LOCALIZATION
-  # ~~~~~
+    project {
 
-  localization {
+      # Enter the list of countries as ISO country codes (e.g. DE). Default country first.
+      # If undefined or empty list, the application will use the countries defined in your commercetools project.
+      # countries = ["DE", "US", "GB", "AT", "CH"]
 
-    # Enter the list of countries as ISO country codes (e.g. DE). Default country first.
-    # If undefined or empty list, the application will use the countries defined in your commercetools project.
-    # countries = ["DE", "US", "GB", "AT", "CH"]
+      # Enter the list of currencies as ISO currency codes (e.g. EUR). Default currency first.
+      # If undefined or empty list, the application will use the currencies defined in your commercetools project.
+      # currencies = ["EUR", "USD"]
 
-    # Enter the list of currencies as ISO currency codes (e.g. EUR). Default currency first.
-    # If undefined or empty list, the application will use the currencies defined in your commercetools project.
-    # currencies = ["EUR", "USD"]
-
-    # Enter the list of languages as language tags (e.g. en-UK). Default language first.
-    # If undefined or empty list, the application will use the languages defined in your commercetools project.
-    # languages = ["en", "de", "es"]
+      # Enter the list of languages as language tags (e.g. en-UK). Default language first.
+      # If undefined or empty list, the application will use the languages defined in your commercetools project.
+      # languages = ["en", "de", "es"]
+    }
   }
 }
 
@@ -413,4 +412,5 @@ play = {
 
   modules.enabled += "com.commercetools.sunrise.framework.injection.SunriseModule"
   modules.enabled += "com.commercetools.sunrise.ctp.project.ProjectModule"
+  modules.enabled += "com.commercetools.sunrise.ctp.project.LocalizationModule"
 }
